# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth API",
        "version": "0.1"
    },
    "paths": {
        "/api/signup": {
            "post": {
                "summary": "Register a user",
                "operationId": "65b7ef1b1e076c033bf91055f53bd9a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "login": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "summary": "Login a user",
                "operationId": "b5cee56757c3c79c5b5f115fdc33f460",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "login": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}